refactor(components): extract BottomSheet rail config and panel containers (P2-2)
Move the rail item definitions, snap points and panel-classification sets
into railConfig.ts, the dynamic panel imports into panelRegistry.tsx, and
the desktop/mobile panel renderers into PanelContainers.tsx. BottomSheet.tsx
drops from 1,239 to 869 lines with the same code-splitting boundaries.
