+++ src/components/BottomSheet/BottomSheet.tsx
@@ -1,23 +1,13 @@
 "use client";
-
 import {
   Accessibility,
-  AlertTriangle,
-  Bookmark,
   BotMessageSquare,
-  Bus,
   ChevronLeft,
   ChevronRight,
-  CircleParking,
-  Cloud,
-  Heart,
   Menu,
-  Navigation,
-  Search,
   X,
 } from "lucide-react";
 import { AnimatePresence, motion } from "motion/react";
-import dynamic from "next/dynamic";
 import Image from "next/image";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { createPortal } from "react-dom";
@@ -32,177 +22,19 @@ import useNavStore from "@/stores/useNavStore";
 import useOnboardingStore from "@/stores/useOnboardingStore";
 import HomeContent from "./HomeContent";
 import {
-  A11yPanelSkeleton,
-  BusPanelSkeleton,
-  ChatSkeleton,
-  EnvironmentSkeleton,
-  HazardReportSkeleton,
-  NavigationSkeleton,
-  ParkingPanelSkeleton,
-  PlaceSkeleton,
-  RouteContentSkeleton,
-  RoutePlanSkeleton,
-  SavedPlacesSkeleton,
-  StationDetailSkeleton,
-  WelfareSkeleton,
-} from "./PanelSkeletons";
-
-const AIChatBot = dynamic(() => import("@/components/AIChatBot"), {
-  loading: () => <ChatSkeleton />,
-  ssr: false,
-});
-const ExitNavDialog = dynamic(
-  () => import("@/components/Navigation/ExitNavDialog"),
-  { ssr: false },
-);
-const A11yPanel = dynamic(() => import("./A11yPanel"), {
-  loading: () => <A11yPanelSkeleton />,
-  ssr: false,
-});
-const BusPanel = dynamic(() => import("./BusPanel"), {
-  loading: () => <BusPanelSkeleton />,
-  ssr: false,
-});
-const EnvironmentPanel = dynamic(() => import("./EnvironmentPanel"), {
-  loading: () => <EnvironmentSkeleton />,
-  ssr: false,
-});
-const HazardReportPanel = dynamic(() => import("./HazardReportPanel"), {
-  loading: () => <HazardReportSkeleton />,
-  ssr: false,
-});
-const ParkingPanel = dynamic(() => import("./ParkingPanel"), {
-  loading: () => <ParkingPanelSkeleton />,
-  ssr: false,
-});
-const SavedPlacesPanel = dynamic(() => import("./SavedPlacesPanel"), {
-  loading: () => <SavedPlacesSkeleton />,
-  ssr: false,
-});
-const WelfarePanel = dynamic(() => import("./WelfarePanel"), {
-  loading: () => <WelfareSkeleton />,
-  ssr: false,
-});
-const PlaceContent = dynamic(() => import("./PlaceContent"), {
-  loading: () => <PlaceSkeleton />,
-  ssr: false,
-});
-const RoutePlanContent = dynamic(() => import("./RoutePlanContent"), {
-  loading: () => <RoutePlanSkeleton />,
-  ssr: false,
-});
-const RouteContent = dynamic(() => import("./RouteContent"), {
-  loading: () => <RouteContentSkeleton />,
-  ssr: false,
-});
-const NavigationContent = dynamic(() => import("./NavigationContent"), {
-  loading: () => <NavigationSkeleton />,
-  ssr: false,
-});
-const StationDetailContent = dynamic(() => import("./StationDetailContent"), {
-  loading: () => <StationDetailSkeleton />,
-  ssr: false,
-});
-
-const SNAP_POINTS = {
-  peek: 0.12,
-  half: 0.38,
-  full: 0.92,
-};
-
-// --- Rail icon definitions ---
-interface RailItem {
-  id: RailPanel;
-  Icon: React.ComponentType<{ className?: string }>;
-  labelKey: string;
-  color: string;
-}
-
-// §6.2 revision: "route" is dropped from the rail entirely (not even into
-// "更多") — HomeContent already has its own "規劃路線" entry right below the
-// search bar, so a second route-planning door on the rail duplicated that
-// responsibility instead of adding one. Only the two lowest-frequency
-// actions (hazard report, welfare directory) move into RAIL_MORE_ITEMS;
-// everything else people reach for while looking at the map stays one tap
-// away on the visible rail.
-const RAIL_ITEMS: RailItem[] = [
-  { id: "search", Icon: Search, labelKey: "railSearch", color: "text-primary" },
-  {
-    id: "a11y",
-    Icon: Accessibility,
-    labelKey: "railA11y",
-    color: "text-accessibility",
-  },
-  { id: "bus", Icon: Bus, labelKey: "railBus", color: "text-emerald-600" },
-  {
-    id: "parking",
-    Icon: CircleParking,
-    labelKey: "railParking",
-    color: "text-indigo-500",
-  },
-  {
-    id: "saved",
-    Icon: Bookmark,
-    labelKey: "savedPlaces",
-    color: "text-amber-500",
-  },
-  {
-    id: "environment",
-    Icon: Cloud,
-    labelKey: "environment",
-    color: "text-sky-500",
-  },
-];
-
-const RAIL_MORE_ITEMS: RailItem[] = [
-  {
-    id: "hazard",
-    Icon: AlertTriangle,
-    labelKey: "reportHazard",
-    color: "text-amber-500",
-  },
-  { id: "welfare", Icon: Heart, labelKey: "welfare", color: "text-rose-500" },
-];
-
-// --- Mode-driven panels: shown via sheetMode, override rail panels ---
-const MODE_PANELS = new Set([
-  "place",
-  "plan",
-  "route",
-  "navigation",
-  "station",
-]);
-
-// Mode panels whose content component renders its own back/close + title
-// header (PlaceContent, RoutePlanContent, RouteContent, StationDetailContent
-// each have a "{/* Header */}" block at the top of their JSX) — the shared
-// desktop chrome header below must stay hidden for these, exactly like the
-// mobile header already does via its own sheetMode check, or the two stack
-// into a visible double header. "navigation" is deliberately excluded:
-// NavigationContent has no header of its own, so the shared chrome one is
-// the only one and must stay.
-const MODE_PANELS_WITH_OWN_HEADER = new Set([
-  "place",
-  "plan",
-  "route",
-  "station",
-]);
-
-// Rail panels that render their own content component. "search" and "none"
-// both fall through to HomeContent, and "route" is only ever a leftover from
-// RoutePreviewHydrator (the rail routes it through sheetMode="plan" instead),
-// so none of those three count as "a sub-panel is showing". RailPanelOrHome
-// reads the same set, so "is a panel open" can never disagree with what's
-// actually rendered.
-const RAIL_CONTENT_PANELS = new Set<RailPanel>([
-  "a11y",
-  "bus",
-  "parking",
-  "saved",
-  "environment",
-  "hazard",
-  "welfare",
-]);
+  DesktopPanelContent,
+  MobileSheetContent,
+  PanelTitle,
+} from "./PanelContainers";
+import { AIChatBot, ExitNavDialog } from "./panelRegistry";
+import {
+  MODE_PANELS,
+  MODE_PANELS_WITH_OWN_HEADER,
+  RAIL_CONTENT_PANELS,
+  RAIL_ITEMS,
+  RAIL_MORE_ITEMS,
+  SNAP_POINTS,
+} from "./railConfig";
 
 export default function BottomSheet() {
   const { t } = useAppTranslation();
@@ -1034,205 +866,4 @@ export default function BottomSheet() {
       {homeSlot && createPortal(<HomeContent />, homeSlot)}
     </>
   );
-}
-
-// --- Panel title based on active panel ---
-function PanelTitle() {
-  const { t } = useAppTranslation();
-  const { sheetMode, activeRailPanel } = useMapStore(
-    useShallow((s) => ({
-      sheetMode: s.sheetMode,
-      activeRailPanel: s.activeRailPanel,
-    })),
-  );
-
-  const modePanelActive = MODE_PANELS.has(sheetMode);
-  if (modePanelActive) {
-    switch (sheetMode) {
-      case "place":
-        return <>{t("title")}</>;
-      case "plan":
-        return (
-          <>
-            <Navigation className="h-4 w-4 text-blue-500" /> {t("planRoute")}
-          </>
-        );
-      case "route":
-        return (
-          <>
-            <Navigation className="h-4 w-4 text-blue-500" /> {t("route")}
-          </>
-        );
-      case "navigation":
-        return (
-          <>
-            <Navigation className="h-4 w-4 text-blue-500" />{" "}
-            {t("navInstructions")}
-          </>
-        );
-      case "station":
-        return (
-          <>
-            <Accessibility className="h-4 w-4 text-accessibility" />{" "}
-            {t("stationDetail")}
-          </>
-        );
-    }
-  }
-
-  switch (activeRailPanel) {
-    case "search":
-      return (
-        <>
-          <Search className="h-4 w-4 text-primary" /> {t("railSearch")}
-        </>
-      );
-    case "a11y":
-      return (
-        <>
-          <Accessibility className="h-4 w-4 text-accessibility" />{" "}
-          {t("accessibleTitle")}
-        </>
-      );
-    case "bus":
-      return (
-        <>
-          <Bus className="h-4 w-4 text-emerald-600" /> {t("busInfo")}
-        </>
-      );
-    case "parking":
-      return (
-        <>
-          <CircleParking className="h-4 w-4 text-indigo-500" /> {t("parking")}
-        </>
-      );
-    case "saved":
-      return (
-        <>
-          <Bookmark className="h-4 w-4 text-amber-500" /> {t("savedPlaces")}
-        </>
-      );
-    case "environment":
-      return (
-        <>
-          <Cloud className="h-4 w-4 text-sky-500" /> {t("environment")}
-        </>
-      );
-    case "hazard":
-      return (
-        <>
-          <AlertTriangle className="h-4 w-4 text-amber-500" />{" "}
-          {t("reportHazard")}
-        </>
-      );
-    case "welfare":
-      return (
-        <>
-          <Heart className="h-4 w-4 text-rose-500" /> {t("welfare")}
-        </>
-      );
-    default:
-      return <>{t("title")}</>;
-  }
-}
-
-// --- Single source of truth for "which rail panel is showing" — shared by
-// both desktop and mobile so a quick-action click has exactly one code path
-// to render through, instead of desktop reading this switch while mobile
-// fell back to HomeContent's own local subPanel copy of the same logic. ---
-function RailPanelOrHome({
-  onHomeSlotRef,
-}: {
-  onHomeSlotRef: (el: HTMLDivElement | null) => void;
-}) {
-  const { activeRailPanel, setActiveRailPanel } = useMapStore(
-    useShallow((s) => ({
-      activeRailPanel: s.activeRailPanel,
-      setActiveRailPanel: s.setActiveRailPanel,
-    })),
-  );
-  // Back out to the search home view, matching the header's back button —
-  // "none" would close the whole panel on desktop and merely re-render the
-  // same HomeContent on mobile.
-  const closePanel = () => setActiveRailPanel("search");
-
-  switch (activeRailPanel) {
-    case "a11y":
-      return <A11yPanel onClose={closePanel} hideHeader />;
-    case "bus":
-      return <BusPanel onClose={closePanel} hideHeader />;
-    case "parking":
-      return <ParkingPanel onClose={closePanel} hideHeader />;
-    case "saved":
-      return <SavedPlacesPanel onClose={closePanel} hideHeader />;
-    case "environment":
-      return <EnvironmentPanel onClose={closePanel} hideHeader />;
-    case "hazard":
-      return <HazardReportPanel onClose={closePanel} hideHeader />;
-    case "welfare":
-      return <WelfarePanel onClose={closePanel} hideHeader />;
-    // "search" / "none" are the home view itself, and "route" only ever lands
-    // here as a leftover from RoutePreviewHydrator once sheetMode has already
-    // gone back to "home" — all three mean "no sub-panel". Keep this list in
-    // sync with RAIL_CONTENT_PANELS. This marker div is a portal *target*,
-    // not `HomeContent` itself — see the single real mount in `BottomSheet`.
-    default:
-      return <div ref={onHomeSlotRef} className="contents" />;
-  }
-}
-
-// --- Desktop panel content switcher ---
-function DesktopPanelContent({
-  onHomeSlotRef,
-}: {
-  onHomeSlotRef: (el: HTMLDivElement | null) => void;
-}) {
-  const { sheetMode } = useMapStore(
-    useShallow((s) => ({ sheetMode: s.sheetMode })),
-  );
-
-  const modePanelActive = MODE_PANELS.has(sheetMode);
-  if (modePanelActive) {
-    switch (sheetMode) {
-      case "place":
-        return <PlaceContent />;
-      case "plan":
-        return <RoutePlanContent />;
-      case "route":
-        return <RouteContent />;
-      case "navigation":
-        return <NavigationContent />;
-      case "station":
-        return <StationDetailContent />;
-    }
-  }
-
-  return <RailPanelOrHome onHomeSlotRef={onHomeSlotRef} />;
-}
-
-// --- Mobile sheet content ---
-function MobileSheetContent({
-  onHomeSlotRef,
-}: {
-  onHomeSlotRef: (el: HTMLDivElement | null) => void;
-}) {
-  const { sheetMode } = useMapStore(
-    useShallow((s) => ({ sheetMode: s.sheetMode })),
-  );
-  switch (sheetMode) {
-    case "home":
-      return <RailPanelOrHome onHomeSlotRef={onHomeSlotRef} />;
-    case "place":
-      return <PlaceContent />;
-    case "plan":
-      return <RoutePlanContent />;
-    case "route":
-      return <RouteContent />;
-    case "navigation":
-      return <NavigationContent />;
-    case "station":
-      return <StationDetailContent />;
-    default:
-      return <RailPanelOrHome onHomeSlotRef={onHomeSlotRef} />;
-  }
 }
